requirements.txt
opencv-python
pandas
xlsxwriter
python-dotenv
